feat: Migrate to CPR SDK
The data access library has been deprecated.

It looks like this package isn't used directly in here.
It is used in https://github.com/climatepolicyradar/navigator-document-parser
[1][2], through a transitive dependency on this package.

We don't lock the version locally, and our Poetry lock files are thus
out-of-sync with this version. Relax it until we have a better approach.

Don't try and install twice too.

[1] https://github.com/climatepolicyradar/navigator-document-parser/blob/572fd0217abfdac81c83afcf61dbfdfcf5369ae1/poetry.lock#L273
[2] https://github.com/climatepolicyradar/navigator-document-parser/blob/572fd0217abfdac81c83afcf61dbfdfcf5369ae1/src/base.py#L6

test: Align test command